# swagger.yaml
basePath: /api
definitions:
...
  database.CreateCurrencyParams:
    properties:
      code:
        type: string
      description:
        type: string
    type: object
...
  server.CreateExchangerRequest:
    properties:
      description:
        type: string
      in_currency:
        type: string
      inmin:
        type: number
      out_currency:
        type: string
      payment_verification:
        type: boolean
    type: object
...
  server.RemoveCurrencyRequest:
    properties:
      code:
        type: string
    type: object
  server.RemoveExchangerRequest:
    properties:
      id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: Swap Support
    url: http://github.com/d1nch8g/swap
  description: Swap exchange api.
  license:
    name: MIT
    url: https://github.com/d1nch8g/swap/src/branch/main/LICENSE
  termsOfService: http://github.com/d1nch8g/swap
  title: Swap exchange api
  version: "1.0"
paths:
...
  /admin/check-if-admin:
    delete:
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Check if user is an admin
...
  /admin/create-currency:
    post:
      parameters:
      - description: Create currency params
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/database.CreateCurrencyParams'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create new currency in exchanger
  /admin/create-exchanger:
    post:
      parameters:
      - description: Create exchanger parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.CreateExchangerRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create new exchanger with provided currencies
...
  /admin/remove-currency:
    delete:
      parameters:
      - description: Remove currency parameter
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.RemoveCurrencyRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Remove currency from currency list
  /admin/remove-exchanger:
    delete:
      parameters:
      - description: Remove exchanger parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.RemoveExchangerRequest'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Remove existing exchanger from API
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Token authorization for internal operations
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"